Add `npx` to the prepare step.
This is needed for integrators that will use `npm install` after
cloning. The lone `husky` command assumes it's globally installed, and
the install will fail if it's not.

`husky` is used to prevent potential license key leaks, and it should
not bleed this details to the testers. This is the fastest workaround.